gobackend: fix FusedSDPA accumulation and mask validation; add Capabilities.DynamicDimDType (#39)
## Summary

This PR addresses two main areas:

### 1. FusedSDPA Accumulation Fix & Mask/Bias Validation (`gobackend`)
- **Fixed accumulation in AVX2 & AVX-512 kernels**: In
`internal/gobackend/fusedops/avx2/sdpa.go` and
`internal/gobackend/fusedops/avx512/sdpa.go`, fixed vector fused
multiply-add calculation from `out.MulAdd(vW, v)` (which evaluated to
`out * vW + v`) to `vW.MulAdd(v, out)` (evaluating to `vW * v + out`).
This resolves incorrect weighted value accumulation across head
dimensions (32, 64, 128, and general).
- **Mask & Bias Shape Validation**: Added shape compatibility checks
(`isSDPABatchOrHeadCompatible` and `isSDPASeqDimCompatible`) in
`internal/gobackend/fusedops/sdpa.go` to return
`compute.ErrNotImplemented` for unsupported broadcasting across sequence
dimensions or mismatched mask/bias layouts.
- **Unit & Backend Tests**: Added AVX2 and AVX-512 unit tests verifying
SDPA kernels, and added test cases in `support/backendtest/fusedops.go`.

### 2. Dynamic Dimension DType Support (`Capabilities.DynamicDimDType`)
- **Added `DynamicDimDType`**: Added `Capabilities.DynamicDimDType` and
helper `DynamicDimDTypeOrDefault()` (defaulting to `dtypes.Int64`).
- **Go Backend Integration**: Updated `DynamicDimensionSize` and
`DynamicShape` in `gobackend` to return `dtypes.Int64`.
- **Documentation & Tests**: Updated docstrings in `ops.go` and
`ops_dynamic.go`, and updated unit tests in `gobackend/dynamic_test.go`.

+++ internal/gobackend/ops/dynamic.go
@@ -29,7 +29,7 @@ func DynamicDimensionSize(f *gobackend.Function, operandValue compute.Value, axi
 		return nil, errors.Errorf("DynamicDimensionSize: axis %d out of bounds for rank %d", axis, operand.Shape.Rank())
 	}
 	opType := compute.OpTypeDynamicDimensionSize
-	outputShape := shapes.Make(dtypes.Int32)
+	outputShape := shapes.Make(dtypes.Int64)
 	node, _ := f.GetOrCreateNode(opType, outputShape, inputs, axis)
 	return node, nil
 }
@@ -65,7 +65,7 @@ func DynamicShape(f *gobackend.Function, operandValue compute.Value) (compute.Va
 	}
 	operand := inputs[0]
 	opType := compute.OpTypeDynamicShape
-	outputShape := shapes.Make(dtypes.Int32, operand.Shape.Rank())
+	outputShape := shapes.Make(dtypes.Int64, operand.Shape.Rank())
 	node, _ := f.GetOrCreateNode(opType, outputShape, inputs, nil)
 	return node, nil
 }
